--- Assets/GeneralBossF.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 074aca1be8abc6b4e9ad2840ec7b5265
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
