+++ Assets/Scripts/Combat/EnemyHealth.cs
@@ -8,11 +8,22 @@ public class EnemyHealth : MonoBehaviour
     public float MaxHealth = 100.0f;
 
     [HideInInspector]
+    public float burning = 0.0f;
+
+    [HideInInspector]
     public float currentHealth;
 
     private void Start()
     {
         currentHealth = MaxHealth;
+    }
+    private void Update()
+    {
+        if (burning > 0.0f)
+        {
+            burning -= Time.deltaTime;
+            TakeDamage(10f * Time.deltaTime);
+        }
     }
     public void TakeDamage(float damage)
     {
