+++ Assets/Scripts/General.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f0cb464287882d540a49d39b57977182
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
