+++ Assets/Scripts/General/BossDialogue.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eda27abc4c2a827458cff548bdff4788
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
