+++ Assets/Scripts/General/ReleaseGuard.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d187dd9755f54204eab9bd556e9a1458
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
