+++ ProjectSettings/TagManager.asset
@@ -11,6 +11,7 @@ TagManager:
   - Debug
   - Enemy
   - Wall
+  - Interactable
   layers:
   - Default
   - TransparentFX
