+++ UserSettings/EditorUserSettings.asset
@@ -9,22 +9,28 @@ EditorUserSettings:
       value: 5601060303560b5d0c575d7342775e44404e4f79297175317e2f4c36b5e5663a
       flags: 0
     RecentlyUsedSceneGuid-1:
-      value: 0100015e04020c0a0c5a5e7a497b0e4415151a2c7c7d7f612c2b1f65e4e5633a
-      flags: 0
-    RecentlyUsedSceneGuid-2:
       value: 5500010701010b5d0f0c5c7515740b44444f412e7a7c71632f7a4461e7b5303a
       flags: 0
-    RecentlyUsedSceneGuid-3:
+    RecentlyUsedSceneGuid-2:
       value: 5057575e04560d5f5c5d097413740c4414161e2f2a7e7365792d4b35e6e2373c
       flags: 0
-    RecentlyUsedSceneGuid-4:
+    RecentlyUsedSceneGuid-3:
       value: 5b030605550c0b580c5b5f7145715e4443164d29292d25692e2a4463e7b4303c
       flags: 0
+    RecentlyUsedSceneGuid-4:
+      value: 0109010754045e5e545d092312710a44424e4879747870627f2f4d63b5e13639
+      flags: 0
     RecentlyUsedSceneGuid-5:
-      value: 55535055515659595f575c2749200744474e4f2e7a782765747a4b62e6b9616c
+      value: 07000100520208590b0b5b7248705b4447151e7979702561797f4931bae2313b
       flags: 0
     RecentlyUsedSceneGuid-6:
+      value: 0100015e04020c0a0c5a5e7a497b0e4415151a2c7c7d7f612c2b1f65e4e5633a
+      flags: 0
+    RecentlyUsedSceneGuid-7:
       value: 500404525402505a0e0f0f7340210944464f1b7e28297e602c284564e3e1303a
+      flags: 0
+    RecentlyUsedSceneGuid-8:
+      value: 55535055515659595f575c2749200744474e4f2e7a782765747a4b62e6b9616c
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
